Remove PpmCanvas
